Started adding inertia cancelling when halting, to prevent overrunning the desired location
Taking size into account when planning, so never plan to move right/left if moving that way would cause Mario to head-butt a brick due to his size
Added MovementInstruction - which can be used to further fine-tune Mario's movements, particularly wrt inertia (currently not incorporated into the Agent)

+++ src/ch/idsia/agents/controllers/kbarrett/MapSquareWrapper.java
@@ -10,7 +10,7 @@ public class MapSquareWrapper {
 	private MapSquare mapSquare;
 	private MapSquareWrapper parent;
 	private Direction direction;
-	private int g = -1;
+	private float g = -1;
 	private int h = -1;
 	private int levelInJump = 0;
 	private int widthOfJump = 0;
@@ -122,11 +122,11 @@ public class MapSquareWrapper {
 			{
 				this.h = h;
 			}
-			public void setG(int g)
+			public void setG(float g)
 			{
 				this.g = g;
 			}
-			public int getG()
+			public float getG()
 			{
 				return g;
 			}
